requirements.txt
python-dotenv
pytesseract
pdf2image
Pillow
langchain
langchain-groq
faiss-cpu
sentence-transformers
langchain-community
fitz
pypdf
pymupdf